IN_main_code - moved filter to helper
